requirements.txt
amqp==5.2.0
asgiref==3.8.1
attr==0.3.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.1.0
billiard==4.2.0
build==0.9.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
daphne==4.1.2
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-rest-auth==5.1.0
Django==4.2.11
django-appconf==1.0.6
django-braces==1.15.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-compressor==4.4
django-cors-headers==4.3.1
django-cors-middleware==1.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-environ==0.11.2
django-extensions==1.7.1
django-filter==23.5
django-import-export==4.0.1
django-js-asset==2.2.0
django-model-utils==4.5.0
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-sass-processor==1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.19.0
et-xmlfile==1.1.0
filelock==3.14.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
imageio==2.33.1
imageio-ffmpeg==0.4.9
import-export==0.3.1
importlib==1.0.4
importlib-resources==5.10.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jwcrypto==1.5.6
kombu==5.3.7
libsass==0.23.0
Markdown==3.5.1
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.3
packaging==21.3
pandas==2.2.2
pep517==0.13.0
phonenumbers==8.13.31
pillow==10.3.0
pip-chill==1.0.3
pip-tools==6.9.0
platformdirs==4.2.1
proglog==0.1.10
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.1
rsa==4.9
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
stripe==9.10.0
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
wheel==0.43.0
yarg==0.1.9
zope.interface==6.3
